changes positon folder layout
