+++ src/common/layout.tsx
@@ -2,14 +2,12 @@ import { TOKEN } from '@constance/cookie';
 import { getCookies } from '@cookies';
 import { LayoutProps, Layout } from 'antd';
 import { Navigate, Outlet } from 'react-router-dom';
-import { MenuNavigator } from '../components/layout/main-menu';
-import { useToggle } from '../hooks/useToggle';
 import { ContentApp } from './app-content';
+import { SiderApp } from './sider';
 
-const { Sider } = Layout;
 
 export const LayoutApp: React.FC<LayoutProps> = () => {
-  const [collapsed, onCollapse] = useToggle();
+
   const token = getCookies(TOKEN);
   if (!token[0]) return <Navigate to={'/login'} />;
   // @ts-ignore
@@ -17,16 +15,7 @@ export const LayoutApp: React.FC<LayoutProps> = () => {
 
   return (
     <Layout style={{ minHeight: '100vh' }}>
-      <Sider
-        className='layout-sider'
-        trigger={null}
-        theme='light'
-        width={250}
-        collapsible
-        collapsed={collapsed}
-      >
-        <MenuNavigator onCollapse={onCollapse} collapsed={collapsed} />
-      </Sider>
+      <SiderApp />
 
       <ContentApp>
         <Outlet />
