+++ src/stylesheets/pages/_admin.scss
@@ -1,5 +1,5 @@
 @import '../base/variables';
-.wrapper-title-admin-page {
+.wrapper-title-page {
     display: flex;
     justify-content: space-between;
     padding: 10px;
